20200812 Greek Update commit:
Updates on Amstelvar-Roman-opsz14-wght500-wdth125 and Amstelvar-Roman-opsz14-wght900-wdth100 for the  extrapolation Amstelvar-Roman-opsz14-wght900-wdth1125.

Green:

Lower case sigma: you are right it was wrong, I had to updated both the 14 400 125 and 14 900 100

Red:

beta: the junction looks too thin I think, so I checked the 14 900 100, and made it heavier by 50% and a tiny bit heavier in the 14 400 125. Please extrapolate again the beta.

delta: looked too narrow so I updated the 14 900 100 and 14 400 125. Please extrapolate delta again.

xi and zeta: both seemed narrow in 14 900 125 so I updated the 14 900 100 and 14 400 125. Please extrapolate again both zeta and xi.

tau: also looked too narrow so I updated the 14 900 100 and 14 400 125. Please extrapolate again both tau.

upsilon and psi: the contrast looked a bit weird in the extrapolated 14 900 100, so I reworked in both 14 900 100 and 14 400 125. Please extrapolate again both.

omega: it looked a bit thin, especially on the base so I updated the 14 900 100. Please extrapolate again omega.